% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplotDIV.R
\name{boxplotDIV}
\alias{boxplotDIV}
\title{Diversity indexes by cluster}
\usage{
boxplotDIV(
  data,
  somC,
  labels = c("Ind/10cm2", "Nº gêneros", "Shannon-Winner", "Simpson", "Jaccard")
)
}
\arguments{
\item{data}{data.frame containing the variable to be plotted.}

\item{somC}{object resulted from \code{\link{wrapSOM}} function.}
}
\value{
A graphical Output
}
\description{
Generates graphical outputs
}
\examples{
boxplotDIV(nema_PCRBS, somC)
}
\author{
Danilo Candido Vieira
}
