% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecoMLreport.R
\name{ecoMLreport}
\alias{ecoMLreport}
\title{Generate Report}
\usage{
ecoMLreport(som.graphics, rf.graphics, output_file = "ReportEcoML", output_dir)
}
\arguments{
\item{som.graphics}{- the list of graphics resulted from the \code{\link{exploreSOM}} function}

\item{rf.graphics}{- the list of graphics resulted from the \code{\link{exploreRF}} function}

\item{output_dir}{- The output directory for the rendered output_file. If is missing the file will be saved in the working directory.}
}
\value{
The function reders a HTML in the specified output directory
}
\description{
#' Generates a report for the analyses
}
