% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exploreSOM.R
\name{exploreSOM}
\alias{exploreSOM}
\title{Exploring the SOMwrap results}
\usage{
exploreSOM(
  somC,
  coords,
  spd,
  layer1 = NULL,
  layer2 = NULL,
  layer3 = NULL,
  div = T,
  npic = 5,
  indicate = "cor",
  scaling = 1.4,
  res = 2000,
  crs.info = "+proj=longlat",
  ...
)
}
\arguments{
\item{somC}{object resulted from \code{\link{wrapSOM}} function.}

\item{coords}{data.frame containing the Spatial coordinates}

\item{spd}{An object of class \link[sp:SpatialPolygons]{SpatialPolygonsDataFrame}.It will be used to interpolate results.}

\item{layer1}{Optional, polygon. An object of class \code{\link[sp:SpatialPolygons]{SpatialPolygons}}}

\item{layer2}{Optional, points. An object of class \code{\link[sp:SpatialPolygons]{SpatialPointsDataFrame}}}

\item{layer3}{Optional, points. An object of class \code{\link[sp:SpatialPolygons]{SpatialPointsDataFrame}}}

\item{div}{Logical. Explore diversity indexes?}

\item{npic}{Integer. Number of indicator variables to plot}

\item{indicate}{Defaults to \code{indicate="cor"}. See \code{\link{somInd}}}

\item{scaling}{Integer. Scaling factor for the arrows when plotting \code{\link{somInd}}}

\item{res}{Integer. Interpolation resolution}

\item{crs.info}{string with CRS information. More details in \code{sp::\link[sp:CRS-class]{CRS}}}

\item{main}{Optional, string. Overhall title of the plot}

\item{leg}{Optional, string.Title of the legend}
}
\value{
A graphical output

Returns a list with graphical outputs from the \code{\link{wrapSOM}} analyses;
}
\description{
Creates a list of graphical outputs using a \code{\link{wrapSOM}} object
}
\examples{
exploreSOM(somC,coords_PCRBS,spd_PCRBS)
}
\author{
Danilo Candido Vieira
}
