% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getGroups.R
\name{getGroups}
\alias{getGroups}
\title{Number of clusters}
\usage{
getGroups(data, seed = NULL, group_method = "kmeans")
}
\arguments{
\item{data}{- data.frame}

\item{seed}{Integer. The random number generator.}

\item{group_method}{the cluster analysis method to be used. This should be one of: "ward.D", "ward.D2", "single", "complete", "average", "mcquitty", "median", "centroid", "kmeans". More details in \code{\link[NbClust:NbClust]{Nbclust}} help page.}
}
\value{
The function retunrs a list containing (1) the best number of cluster (the number with the highest number of votes) and (2) the counting table of votes.
}
\description{
Compute the best number of clusters according to different indexes in \code{NbClust::\link[NbClust]{NbClust}} and defines the winning number as the one with the highest number of votes.
}
\examples{
getGroups(nema_PCRBS)
}
