% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapCells.R
\name{mapCells}
\alias{mapCells}
\title{Plot the heatmap  throung a loop of variables}
\usage{
mapCells(m, pic)
}
\arguments{
\item{m}{an object of class \code{'kohonen'}, resulted from the functions \code{\link[kohonen:supersom]{som}} and \code{\link[kohonen]{supersom}}}

\item{pic}{Integer or string indicating the variables (columns) to plot.}
}
\value{
A graphical output
}
\description{
Plot the heatmap for multiple variables across the kohonen map.
}
\examples{
somC<-SOMwrap(nema_PCRBS,dist=c("BrayCurtis"),groups=NULL, seed=1,n_iterations = 500, xdim=5, ydim=5)
indicators<-rownames(somInd(somC$som))
mapCells(somC$som.model, pic=indicators)
}
