% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapClust.R
\name{mapClust}
\alias{mapClust}
\title{Functions to spatialize clusters resulted from  analysis}
\usage{
mapClust(
  somC,
  coords,
  spd,
  layer1 = NULL,
  layer2 = NULL,
  layer3 = NULL,
  res = 20000,
  main = "",
  hlegend = 0.5,
  crs.info = "+proj=longlat",
  leg = ""
)
}
\arguments{
\item{somC}{Object resulted from \code{\link{wrapSOM}} function.}

\item{coords}{Data.frame containing the Spatial coordinates}

\item{spd}{An object of class \code{\link[sp:SpatialPolygons]{SpatialPolygonsDataFrame}}. It will be used to interpolate results.}

\item{layer1}{Optional, polygon. An object of class \code{\link[sp:SpatialPolygons]{SpatialPolygons}}}

\item{layer2}{Optional, points. An object of class \code{\link[sp:SpatialPolygons]{SpatialPointsDataFrame}}}

\item{layer3}{Optional, points. An object of class \code{\link[sp:SpatialPolygons]{SpatialPointsDataFrame}}}

\item{res}{Integer. Interpolation resolution}

\item{main}{Optional, string. Overhall title of the plot}

\item{crs.info}{string with CRS information. More details in \code{sp::\link[sp:CRS-class]{CRS}}}

\item{leg}{Optional, string. Legend title.}
}
\value{
A graphical output
}
\description{
Spatialize the clusters resulted from the \code{\link{wrapSOM}} analysis
}
\examples{
somC<-wrapSOM(nema_PCRBS,dist=c("BrayCurtis"),groups=NULL, seed=1,n_iterations = 500, xdim=5, ydim=5)
mapClust(somC,coords=coords_PCRBS ,spd=spd_PCRBS,crs.info="+proj=longlat")
}
