% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MapUni.R
\name{mapOne}
\alias{mapOne}
\alias{mapLoop}
\title{Functions to spatialize one variable from a dataset.}
\usage{
mapOne(
  data,
  coords,
  get = 1,
  spd,
  layer1 = NULL,
  layer2 = NULL,
  layer3 = NULL,
  res = 20000,
  main = "",
  leg = "",
  crs.info = "+proj=longlat",
  ...
)

mapLoop(data, coords, pic, ...)
}
\arguments{
\item{data}{- data.frame containing the variable to be plotted.}

\item{coords}{data.frame containing the Spatial coordinates.}

\item{get}{- integer or a string containing the variable name to be plotted.}

\item{spd}{-  An object of clas \code{SpatialPolygonsDataFrame}. It will be used to interpolate results.}

\item{layer1}{Optional, polygon. An object of clas \code{SpatialPolygons}}

\item{layer2}{Optional, points.  An object of class \code{SpatialPointsDataFrame}}

\item{layer3}{Optional, points.  An object of class \code{SpatialPointsDataFrame}}

\item{res}{Integer. Resolution of the interpolations}

\item{main}{Optinal, string. Title of the plot}

\item{leg}{Optinal, string. Legend title of the plot}

\item{crs.info}{string with CRS information. More details in \code{sp::\link[sp:CRS-class]{CRS}}}

\item{...}{Arguments passed to \code{\link[ecoML:mapUni]{mapOne}} and \code{\link[gstat:krige]{idw}} functions.}

\item{pic}{Integer or string indicating the variables (columns) to plot.}
}
\value{
The function \code{mapOne} returns a list containing one single plot, whereas \code{mapLoop} returns a list containing multiple plots.
}
\description{
The function \code{mapOne} spatialize one  variable from a given dataset. The function `mapLoop` spatialize multiple variables.
}
\examples{
mapOne(nema_PCRBS,coords=coords_PCRBS,spd=spd_PCRBS)
mapLoop(nema_PCRBS, coords=coords_PCRBS, pic=1:5,spd=spd_PCRBS)
}
\author{
Danilo Candido Vieira
}
