% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nema.R
\name{nema}
\alias{nema}
\title{Nematodes data from Araca Bay, Sao Sebastiao, Brazil.}
\format{\code{nema} is a list with the following components:
\describe{
  \item{envi}{a matrix with 141 sites and 9 environmental variables: chlorophyll a (mg.m-2), bathymetry (meters), percentage of total organic carbon, percentage of coarse sands (as the sum of pebbles, very coarse, coarse, and medium grains), percentage of fine sand, percentage of very fine sand, mean grain size, and sorting coefficient.}
  \item{comm}{matrix with 141 sites of 194 nematodes species}
}}
\usage{
data(nema)
}
\description{
A total of  37 sites arranged in an irregular grid were surveyed during four sampling campaigns at the
Araçá Bay, southeastern Brazil. A total of 141 samples were collected for analyzing changes in
nematodes assemblages along the environmental gradient of the bay.
}
\references{
\itemize{
  \item Checon, H. H., D. C. Vieira, G. N. Corte, E. C. P. M. Sousa, G. Fonseca, and A. C. Z. Amaral. 2018. Defining soft bottom habitats and potential indicator species as tools for monitoring coastal systems: A case study in a subtropical bay. Ocean & Coastal Management.
  \item Corte, G. N., H. H. Checon, G. Fonseca, D. C. Vieira, F. Gallucci, M. Di Domenico, and A. C. Z. Amaral. 2017. Cross-taxon congruence in benthic communities: Searching for surrogates in marine sediments. Ecological Indicators 78:173–182.
}
}
