% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCM.R
\name{plotCM}
\alias{plotCM}
\title{Confusion Matrix plot}
\usage{
plotCM(rf, somC, main = "", cex.text = 1, line = 2.8)
}
\arguments{
\item{rf}{Random forest model resulted from \code{\link{wrapRF}}}

\item{somC}{object resulted from \code{\link{wrapSOM}} function}

\item{main}{title}

\item{cex.text}{size for text labels}

\item{line}{Integer for adjusting the title}
}
\value{
A graphical ouptut
}
\description{
Plot the confusion matrix resulted from \code{RFwrap} ...
}
\examples{
rf=wrapRF(nema_PCRBS,somC)
plotCM(rf, somC)
}
