% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segRDA-package.R
\docType{package}
\name{segRDA-package}
\alias{segRDA}
\alias{segRDA-package}
\title{segRDA: An R package for performing piecewise redundancy analysis}
\description{
'segRDA' is designed to model non-continuous linear responses of ecological communities to environmental data  through three steps: (1) data ordering (function 'OrdData'), (2) split-moving-window analysis (function 'SMW') and (3) piecewise redundancy analysis (function 'pwRDA').
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DaniloCVieira/segRDA}
  \item Report bugs at \url{https://github.com/DaniloCVieira/segRDA/issues}
}

}
\author{
\strong{Maintainer}: Danilo Candido Vieira \email{vieiradc@yahoo.com.br}

}
\keyword{internal}
