% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim1.R, R/sim2.R, R/sim3.R
\docType{data}
\name{sim1}
\alias{sim1}
\alias{sim2}
\alias{sim3}
\title{Simulated datasets}
\format{Each data set is a list with the following components:
\describe{
  \item{envi}{environmental matrix}
  \item{comm}{community matrix}
}}
\usage{
data(sim1)

data(sim2)

data(sim3)
}
\description{
Simulated datasets for testing \pkg{segRDA} package
}
\keyword{datasets}
