% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somInd.R
\name{somInd}
\alias{somInd}
\title{Indicators from SOM}
\usage{
somInd(
  m,
  npic = 5,
  indicate = c("cor", "var"),
  col.arrow = "gray80",
  scaling = 1.4
)
}
\arguments{
\item{m}{an object of class \code{'kohonen'}, resulted from the functions \code{\link[kohonen:supersom]{som}} and \code{\link[kohonen]{supersom}}.}

\item{npic}{Number of indicators variables to plot}

\item{indicate}{Calculation type. If \code{indicate=="cor"}, the function calculates the correlation of the variables with cells (weigthed by number of instances) and plot \code{npic} variables with the highest correlations, considering the different directions of the map. If \code{indicate=="var"}, the function calculates the variance of each variable weigthed by number of instances, and plot \code{npic} variables with the highest variances along the cell map. Variables with higher variances are those wich induced the strongest contrasts between the cells of the map.}

\item{scaling}{Integer. Scaling factor for the arrows.}

\item{somC}{object resulted from \code{\link{wrapSOM}} function.}

\item{pic}{Integer or string indicating the variables (columns) to plot.}
}
\value{
The function invisibly returns the indicator variables, and their respective measure (settled by the argument \code{indicate})
}
\description{
Plot the indicators variables along the kohonen cell map, with  the arrows indicating the direction and strength of the correlation between variables and cells.
}
\examples{
somC<-wrapSOM(nema_PCRBS,dist=c("BrayCurtis"),groups=NULL, seed=1,n_iterations = 500, xdim=5, ydim=5)
somInd(somC$som)
}
