% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/topoerror.R
\name{topoerror}
\alias{topoerror}
\title{Tophological error from SOM}
\usage{
topoerror(m, type = c("nodedist", "bmu"))
}
\arguments{
\item{m}{an objected of class \code{'kohonen'}, resulted from the functions \code{\link[kohonen:supersom]{som}} and \code{\link[kohonen]{supersom}}.}
}
\value{
The tophological error
}
\description{
- In construction
}
