% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/votesPlot.R
\name{votesPlot}
\alias{votesPlot}
\title{Function to plot the best number of clusters}
\usage{
votesPlot(
  somC,
  col1 = "steelblue",
  col2 = "salmon",
  main = "Votes on the best number of clusters",
  xlab = "Best number of clusters",
  ylab = "counting"
)
}
\arguments{
\item{somC}{Object resulted from \code{SOMWrap} function.}

\item{col1}{Color to highlight the number of clusters with the highest vote count.}

\item{main}{Title of the plot}

\item{xlab}{Label of the axis x}

\item{ylab}{Label of the axis y}
}
\value{
A graphical output
}
\description{
Barplot indicating the best number of clusters.
}
\examples{
somC<-SOMwrap(nema_PCRBS,dist=c("BrayCurtis"),groups=NULL, seed=1,n_iterations = 500, xdim=5, ydim=5)
votesPlot(somC)
}
