% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapRF.R
\name{wrapRF}
\alias{wrapRF}
\title{Wrap for Random Forest}
\usage{
wrapRF(
  data,
  somC,
  ntree = 500,
  seed = NULL,
  prev.idw = F,
  trainControl.args = list(method = "repeatedcv", number = 10, repeats = 20),
  coords = NULL,
  crs.info = "+proj=longlat",
  ...
)
}
\arguments{
\item{data}{numerical data.frame}

\item{somC}{object resulted from \code{\link{wrapSOM}} function}

\item{ntree}{Integer. Number of trees}

\item{seed}{Integer. The random number generator}

\item{prev.idw}{Logical. Should the supervisor use interpolated data?}

\item{trainControl.args}{List with arguments passed to \code{\link[caret:trainControl]{trainControl}}}

\item{crs.info}{string with CRS information. More details in \code{sp::\link[sp:CRS-class]{CRS}}}
}
\value{
A list is returned of class \code{\link[caret]{train}}
}
\description{
Wrap of functions to perform Random Forest analysis with cross validation using the function \code{caret::\link[caret]{trainControl}}
}
\examples{
rf=wrapRF(nema_PCRBS,somC)
}
