% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapSOM.R
\name{wrapSOM}
\alias{wrapSOM}
\title{Wrap for Self-organizing Maps}
\usage{
wrapSOM(
  data,
  dist = c("BrayCurtis"),
  topo = "hexagonal",
  hmethod = "ward.D2",
  groups = NULL,
  seed = NULL,
  n_iterations = 1000,
  xdim = 8,
  ydim = 6,
  scale = F,
  members = NULL,
  ...
)
}
\arguments{
\item{data}{numerical data.frame}

\item{dist}{distance function to be used for the data. Same as the argument \code{dist.fcts} from \code{kohonen::\link[kohonen:supersom]{som}} function, with the add that the wrap also implements the 'BrayCurtis' dissimilarity measure.}

\item{topo}{choose between a hexagonal or rectangular topology.}

\item{hmethod}{described in argument \code{method} from \code{\link[stats]{hclust}}}

\item{groups}{Integer. The number of groups for clustering. If \code{groups=NULL} (default), the function will apply the function \code{\link{getGroups}} to determine the number of groups.}

\item{seed}{Integer. The random number generator}

\item{n_iterations}{Integer. The number of  interactions.}

\item{xdim}{number of cells in dimension X}

\item{ydim}{number of cells in dimension Y}

\item{scale}{Logical. Should the data be scaled?}

\item{members}{described in argument \code{members} from \code{\link[stats]{hclust}}}

\item{...}{Further arguments passed to \code{kohonen::\link[kohonen]{supersom}}}
}
\value{
Returns an list of length 7:
 \enumerate{
  \item \code{..$somC}: vector giving the cluster identification for each sample
  \item \code{..$colhabs}: the color attributed to each cluster
  \item \code{..$som.model}: object of class \code{kohonen}, resulted from \code{kohonen::\link[kohonen]{supersom}}
  \item \code{..$som.hc}: vector giving the cluster identification for each cell in the codebook
  \item \code{..$groups}: number of cluster defined/proposed
  \item \code{..$groups.result}: votes on the best number of clusters
}
}
\description{
Wrap of functions to perform self-organizing maps analysis and determining the best number of clusters. The wrap uses the function \code{kohonen::\link[kohonen]{supersom}}
}
\author{
Danilo Candido Vieira
}
