% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrdData.R
\name{OrdData}
\alias{OrdData}
\title{Data ordering}
\usage{
OrdData(x, y, axis = 1, ...)
}
\arguments{
\item{x}{explanatory matrix;}

\item{y}{community matrix;}

\item{axis}{the RDA axis in which the ordering should be based. Defaults to \code{axis=1}.}

\item{...}{parameters passed to vegan::\code{\link[vegan]{cca}};}
}
\value{
An object of class \code{"ord"}, which is a list consisting of:
\enumerate{
\item{xo}: the ordered explanatory matrix
\item{yo}: the ordered community matrix
\item{x}: the original explanatory matrix
\item{y}: the original community matrix
}
}
\description{
Ordinates both community and explanatory matrices based on the first RDA score.
}
\examples{
data(sim1)
sim1.o<-OrdData(x=sim1$envi, y=sim1$comm)
}
\author{
Danilo Candido Vieira
}
