% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/segRDA-package.R
\docType{package}
\name{segRDA-package}
\alias{segRDA}
\alias{segRDA-package}
\title{segRDA: Modeling Non-Continuous Linear Responses of Ecological Data}
\description{
Tools for modeling non-continuous linear responses of ecological communities to environmental data. The package is straightforward through three steps: (1) data ordering (function OrdData()), (2) split-moving-window analysis (function SMW()) and (3) piecewise redundancy analysis (function pwRDA()). Relevant references include Cornelius and Reynolds (1991) <doi:10.2307/1941559> and Legendre and Legendre (2012, ISBN: 9780444538697).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/DaniloCVieira/segRDA}
  \item Report bugs at \url{https://github.com/DaniloCVieira/segRDA/issues}
}

}
\keyword{internal}
